decoupled vcd timescale from core ref period. core ref period is now deciding machine unit. time manager is now set by the core object with the ref period. modified signal manager to work with ref period. renamed dax.sim.sim to dax.sim.ddb. streamlined sim logging. added tests for dax.sim.time. updated doc accordingly. other minor fixes and improvements.

+++ test/stubs/numpy.pyi
@@ -4,7 +4,7 @@ int64 = int
 
 
 # noinspection PyPep8Naming
-class integer:
+class integer(int):
     ...
